Changed to VQVAE, implemented encoder and decoder classes
